LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/view-timeline-range-expected.txt


PASS View timeline with range as <name> <percent> pair.
PASS View timeline with range and inferred name or offset.
PASS View timeline with range as <name> <px> pair.
FAIL View timeline with range as <name> <percent+px> pair. assert_equals: Effect at the start of the active phase: contain undefined% to contain undefined% expected "0.3" but got "0.7"
FAIL View timeline with range as strings. assert_equals: Effect at the start of the active phase:  to  expected "0.3" but got "0.7"
